+++ ctor/vo_regular_bp_continuator.py
@@ -1,0 +1,3 @@
+from ctor.vo_regular_bp.continuator import VORegularBPContinuator
+
+__all__ = ["VORegularBPContinuator"]